+++ includes/structs.h
@@ -6,7 +6,7 @@
 /*   By: mmaquine <mmaquine@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/02 14:04:28 by mmaquine          #+#    #+#             */
-/*   Updated: 2026/04/01 18:59:46 by mmaquine         ###   ########.fr       */
+/*   Updated: 2026/04/07 15:53:28 by gabrgarc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,16 +35,16 @@ typedef struct s_canva
 
 typedef struct s_point
 {
-	double	x;
-	double	y;
-	double	z;
+	REAL	x;
+	REAL	y;
+	REAL	z;
 }	t_point;
 
 typedef struct s_matrix
 {
 	int		row;
 	int		col;
-	double	**a;
+	REAL	**a;
 }	t_matrix;
 
 typedef struct s_scene_obj
@@ -55,7 +55,7 @@ typedef struct s_scene_obj
 typedef struct s_amb_light
 {
 	t_obj_type	type;
-	double		light_rate;
+	REAL		light_rate;
 	t_color		color;
 }	t_amb_light;
 
@@ -64,14 +64,14 @@ typedef struct s_cam
 	t_obj_type	type;
 	t_point		point;
 	t_point		orient;
-	double		fov;
+	REAL		fov;
 }	t_cam;
 
 typedef struct s_light
 {
 	t_obj_type	type;
 	t_point		coord;
-	double		bright_rate;
+	REAL		bright_rate;
 	t_color		color;
 }	t_light;
 
@@ -80,7 +80,7 @@ typedef struct s_sphere
 	t_obj_type	type;
 	t_point		center;
 	t_color		color;
-	double		diam;
+	REAL		diam;
 }	t_sphere;
 
 typedef struct s_plane
@@ -96,8 +96,8 @@ typedef struct s_cylinder
 	t_obj_type	type;
 	t_point		center;
 	t_point		v_axis;
-	double		diam;
-	double		height;
+	REAL		diam;
+	REAL		height;
 	t_color		color;
 }	t_cylinder;
